- Used vsnprintf in Logger instead of vsprintf to prevent buffer overflows - introduced 5 digit serial numbers that should be assigned to the nodes when flashing the firmware. These should be unique to the also intoruced manufacturer id, that corrensponds to the company identifier assigned by the Bluetooth SID - Enrollment does now work by serial - Added seperators for uart outputs to allow for better parsing.

+++ src/utility/SimpleQueue.cpp
@@ -29,7 +29,7 @@ bool SimpleQueue::Put(u8* data, u32 dataLength)
 	
 	//If the writePointer is ahead (or at the same point) of the read pointer && bufferSpace
 	//at the end is not enough && dataSize at the beginning is enough
-	if(writePointer >= readPointer && bufferEnd - writePointer <= elementSize && readPointer - bufferStart >= elementSize){
+	if(writePointer >= readPointer && (u32)(bufferEnd - writePointer) <= elementSize && (u32)(readPointer - bufferStart) >= elementSize){
 		writePointer = bufferStart;
 		((u32*)writePointer)[0] = 0;
 	}
